move type difference formatting to FormatType
