+++ main/src/ca/uwaterloo/flix/language/ast/Scheme.scala
@@ -17,6 +17,7 @@
 package ca.uwaterloo.flix.language.ast
 
 import ca.uwaterloo.flix.api.Flix
+import ca.uwaterloo.flix.language.debug.{Audience, FormatType}
 import ca.uwaterloo.flix.language.phase.unification.Unification
 import ca.uwaterloo.flix.util.{InternalCompilerException, Result}
 import ca.uwaterloo.flix.util.tc.Show.ShowableSyntax
@@ -136,15 +137,16 @@ object Scheme {
   * Representation of polytypes.
   */
 case class Scheme(quantifiers: List[Type.Var], base: Type) {
+  private implicit val audience: Audience = Audience.External
 
   /**
     * Returns a human readable representation of the polytype.
     */
   override def toString: String = {
     if (quantifiers.isEmpty)
-      base.show
+      FormatType.format(base)
     else
-      s"∀(${quantifiers.map(tvar => tvar.getText.getOrElse(tvar.id)).mkString(", ")}). ${base.show}"
+      s"∀(${quantifiers.map(tvar => tvar.getText.getOrElse(tvar.id)).mkString(", ")}). ${FormatType.format(base)}"
   }
 
 }
